app/requirements.txt
flask==2.2.3
werkzeug==2.2.3
docker==7.1.0
requests==2.32.3
python-dotenv==1.0.1
pip==24.3.1